+++ src/components/Layout/MDXContent.tsx
@@ -15,7 +15,7 @@ import {
 } from "components/Card/CustomNotice";
 import { Locale, PathConfig, FrontMatter, BuildType } from "shared/interface";
 import { useTotalContributors } from "components/Contributors";
-import replaceInternalHref from "utils/anchor";
+import replaceInternalHref from "shared/utils/anchor";
 import { Pre } from "components/MDXComponents/Pre";
 
 export default function MDXContent(props: {
