+++ src/templates/DocTemplate.tsx
@@ -1,18 +1,16 @@
 import * as React from "react";
-import { graphql, Link } from "gatsby";
-import { Trans, useI18next } from "gatsby-plugin-react-i18next";
+import { graphql } from "gatsby";
+import { useI18next } from "gatsby-plugin-react-i18next";
 import clsx from "clsx";
 import Box from "@mui/material/Box";
 import Stack from "@mui/material/Stack";
 
-// import "styles/global.css";
 import "styles/docTemplate.css";
 
 import Layout from "components/Layout";
 import { LeftNavDesktop, LeftNavMobile } from "components/Navigation/LeftNav";
 import MDXContent from "components/Layout/MDXContent";
 import RightNav, { RightNavMobile } from "components/Navigation/RightNav";
-import ScrollToTopBtn from "components/Button/ScrollToTopBtn";
 import {
   TableOfContent,
   PageContext,
@@ -22,7 +20,7 @@ import {
   Locale,
 } from "shared/interface";
 import Seo from "components/Layout/Seo";
-import { getStable, generateUrl } from "utils";
+import { getStable, generateUrl } from "shared/utils";
 import GitCommitInfoCard from "components/Card/GitCommitInfoCard";
 import { FeedbackSection } from "components/Card/FeedbackSection";
 import { FeedbackSurveyCampaign } from "components/Campaign/FeedbackSurvey";
